chore: update example
