+++ packages/disco/example/test/disco_test.dart
@@ -1,6 +1,5 @@
 import 'package:disco/src/disco_internal.dart';
 import 'package:example/main.dart';
-import 'package:example/model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
