+++ packages/disco/pubspec.yaml
@@ -20,5 +20,5 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  very_good_analysis: ^7.0.0
   mockito: ^5.4.5
+  very_good_analysis: ^7.0.0
